Add `traad-rename-advanced'
Add a new method, `traad-rename-advanced' that allows the user to specify some
special options in the rename action.

This is accomplished by changing `traad-rename' so it can take some new,
optional arguments. It can still be called interactively as before.
`traad-rename-advanced' can be used to explicitly prompt for these options.

Options are:

`docstrings' (default t) - Whether to rename instances of the name in docstrings
                           and comments.
`in-hierarchy' (default t) - Whether to rename instances of the name at other
                             positions in the hierarchy.

Note that this changes the default behaviour of `traad-rename'. Before, it would
not rename in comments/docstrings, or elsewhere in the hierarchy.