Add `traad-goto-definition' commands.
Add a command, `traad-goto-definition', that jumps to the definition of the
thing at point. Note that this requires an equivalent commit in the traad
project.

Also adds an asynchronous version of command that blocks until completion,
`traad-goto-definition-synchronous'. This may be useful if the jump needs to be
called programmatically.

(Rope appears to have more than one function for finding the definition. This
uses the function in the `code_assist` module. Another function exists in the
`findit` module, but it is not used here.)