Add findit commands for locating other versions of objects
Interface with the traad-server to allow the user to utilise the findit module
in Rope. This commit primarily adds three user-facing commands (plus backend
logic):

- `traad-find-occurrences' - Display occurrences of the object under point.
                             Includes unsure occurrences.
- `traad-find-implementations' - Display implementations of the object under
                                 point.
- `traad-find-definition' - Display the definition of the object under point.

All commands display their results in a dedicated buffer, with shortcuts to
preview the result and visit it directly.

Note that this also removes the previous, commented-out versions of these
commands. It also requires a similar commit in the original `traad` project.